+++ src/component/space/index.ts
@@ -1,5 +1,3 @@
-export { CategoryButton } from "./CategoryButton";
-export { FieldButton } from "./FieldButton";
 export { CreateSpace } from "./create/CreateSpace";
 export { RetrospectBox } from "./view/RetrospectBox";
 export { SpaceCountView } from "./view/SpaceCountView";
